Initial commit of new testing infrastructure
