+++ server/googleActions.js
@@ -11,7 +11,7 @@ const config = require('nconf')
 
 //let oAuth2Client;
 
-GoogleActions = {
+const GoogleActions = {
 
 	authCheck : (account, res)=>{
 		if(!account || !account.googleId){ // If not signed into Google
