--- tests/basic.test.js
@@ -1,7 +1,0 @@
-const test = require('pico-check');
-
-test('Just setting up a spot for future tests', (t)=>{
-	t.pass();
-});
-
-module.exports = test;
